requirements.txt
Flask
python-dotenv
groq
pytesseract
Pillow
Flask-SQLAlchemy
